requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
confluent-kafka==2.10.0
dnspython==2.7.0
dotenv==0.9.9
passlib==1.7.4
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
python-dotenv==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
